refactor(penrose-tile): penrose-tileにStillSketchを適用, モジュールを別のファイルに外だし
